refactor(Clickable): mv use client code
